+++ src/main/java/ru/project/drivingschool/repository/jpa/JpaCompanyRepository.java
@@ -3,6 +3,6 @@ package ru.project.drivingschool.repository.jpa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.project.drivingschool.model.Company;
 
-public interface JpaCompanyRepository extends JpaRepository<Company, Integer> {
+public interface JpaCompanyRepository extends JpaRepository<Company, Long> {
 
 }
